+++ hw08/src/main/java/ru/otus/hw/models/Book.java
@@ -6,7 +6,6 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
-import ru.otus.hw.dto.Comment;
 
 import java.util.ArrayList;
 import java.util.List;
