+++ hw08/src/main/java/ru/otus/hw/models/Comment.java
@@ -1,0 +1,23 @@
+package ru.otus.hw.models;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.bson.types.ObjectId;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Comment {
+
+    private String id;
+
+    private String text;
+
+    public Comment(String text) {
+        this.id = ObjectId.get().toString();
+        this.text = text;
+    }
+}
